chore: add less support
The next-with-less plugin is a temporary fix to add support for .less files to the nextjs build process, in the same style as the built in css/sass support. It may not be needed in the future. Track the status of this issue: https://github.com/vercel/next.js/pull/23185 .